# GeneratedApiClient/src/StudentSurveySystem.ApiClient/Model/QuestionType.cs
/* 
 * Student survey system API
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System.Runtime.Serialization;
using Newtonsoft.Json;
using Newtonsoft.Json.Converters;

namespace StudentSurveySystem.ApiClient.Model
{
    /// <summary>
    /// Defines QuestionType
    /// </summary>
    [JsonConverter(typeof(StringEnumConverter))]
        public enum QuestionType
    {
        /// <summary>
        /// Enum Text for value: Text
        /// </summary>
        [EnumMember(Value = "Text")]
        Text = 1,
        /// <summary>
        /// Enum SingleSelect for value: SingleSelect
        /// </summary>
        [EnumMember(Value = "SingleSelect")]
        SingleSelect = 2,
        /// <summary>
        /// Enum MultipleSelect for value: MultipleSelect
        /// </summary>
        [EnumMember(Value = "MultipleSelect")]
        MultipleSelect = 3,
        /// <summary>
        /// Enum Numeric for value: Numeric
        /// </summary>
        [EnumMember(Value = "Numeric")]
        Numeric = 4,
        /// <summary>
        /// Enum Date for value: Date
        /// </summary>
        [EnumMember(Value = "Date")]
        Date = 5,
        /// <summary>
        /// Enum Boolean for value: Boolean
        /// </summary>
        [EnumMember(Value = "Boolean")]
        Boolean = 6    }
}