# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Student survey system API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://localhost:5001/api"
    }
  ],
  "paths": {
    "/api/SurveyResponses": {
      "get": {
        "tags": [
          "SurveyResponses"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "",
              "nullable": true
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "count",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SurveyResponseDetailsDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SurveyResponseDetailsDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SurveyResponseDetailsDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "SurveyResponses"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SurveyResponseDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SurveyResponseDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SurveyResponseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SurveyResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurveyResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurveyResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/SurveyResponses/{id}": {
      "get": {
        "tags": [
          "SurveyResponses"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SurveyResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurveyResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurveyResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/SurveyResponses/MyCompleted": {
      "get": {
        "tags": [
          "SurveyResponses"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "",
              "nullable": true
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "count",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SurveyResponseDetailsDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SurveyResponseDetailsDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SurveyResponseDetailsDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Surveys/{id}": {
      "get": {
        "tags": [
          "Surveys"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SurveyDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurveyDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurveyDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Surveys"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SurveyDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SurveyDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SurveyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Surveys"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "QuestionType": {
        "enum": [
          "Text",
          "SingleSelect",
          "MultipleSelect",
          "Numeric",
          "Date",
          "Boolean"
        ],
        "type": "string"
      },
      "AnswerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "value": {
            "type": "string",
            "nullable": true
          },
          "questionText": {
            "type": "string",
            "nullable": true
          },
          "questionType": {
            "$ref": "#/components/schemas/QuestionType"
          },
          "questionId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "SurveyResponseDetailsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "respondentId": {
            "type": "integer",
            "format": "int32"
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnswerDto"
            },
            "nullable": true
          },
          "surveyName": {
            "type": "string",
            "nullable": true
          },
          "respondent": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "creator": {
            "type": "string",
            "nullable": true
          },
          "courseName": {
            "type": "string",
            "nullable": true
          },
          "surveyId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "SurveyResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "respondentId": {
            "type": "integer",
            "format": "int32"
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnswerDto"
            },
            "nullable": true
          },
          "surveyId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "QuestionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "questionText": {
            "type": "string",
            "nullable": true
          },
          "questionType": {
            "$ref": "#/components/schemas/QuestionType"
          },
          "values": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SurveyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "creatorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "courseId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionDto"
            },
            "nullable": true
          },
          "active": {
            "type": "boolean"
          },
          "courseName": {
            "type": "string",
            "nullable": true
          },
          "creatorName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme.",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}